+++ Content.Client/_CE/ZLevels/Roof/CEClientZLevelsRoofSystem.cs
@@ -7,5 +7,4 @@ using Content.Shared._CE.ZLevels.Roof;
 
 namespace Content.Client._CE.ZLevels.Roof;
 
-/// <inheritdoc/>
-public sealed class CEClientRoofSystem : CESharedRoofSystem;
+public sealed class CEClientZLevelsRoofSystem : CESharedZLevelsRoofSystem;
