+++ Content.Server/_CE/ZLevels/Roof/CEZLevelsRoofSystem.cs
@@ -1,0 +1,21 @@
+/*
+ * This file is sublicensed under MIT License
+ * https://github.com/space-wizards/space-station-14/blob/master/LICENSE.TXT
+ */
+
+using Content.Shared._CE.ZLevels.Roof;
+
+namespace Content.Server._CE.ZLevels.Roof;
+
+public sealed partial class CEZLevelsRoofSystem : CESharedZLevelsRoofSystem
+{
+    private readonly HashSet<Vector2i> _roofMap = new();
+
+    public override void Initialize()
+    {
+        base.Initialize();
+
+        InitMaps();
+        InitGrids();
+    }
+}
