+++ Content.Shared/_CE/ZLevels/Roof/CEZLevelRoofComponent.cs
@@ -8,8 +8,8 @@ using Robust.Shared.GameStates;
 namespace Content.Shared._CE.ZLevels.Roof;
 
 /// <summary>
-/// Marks a map for automatic roof management driven by tile changes in the Z-level system.
+/// Marks a mapGrid entity (supports both planetmap + grids) for automatic roof management driven by tile changes in the Z-level system.
 /// Systems use this marker to add, update, or remove roof tiles when the underlying tiles change.
 /// </summary>
 [RegisterComponent, NetworkedComponent]
-public sealed partial class CEZLevelMapRoofComponent : Component;
+public sealed partial class CEZLevelRoofComponent : Component;
