+++ Resources/Prototypes/_CE/Maps/debug.yml
@@ -72,7 +72,7 @@
             - type: SunShadowCycle
             - type: LightCycle
               originalColor: "#ffed75"
-            - type: CEZLevelMapRoof
+            - type: CEZLevelRoof
             - type: SunShadow
             - type: MapLight
               ambientLightColor: "#ffed75"
@@ -102,7 +102,7 @@
     - type: LightCycle
       offset: 490
       originalColor: "#c8fdff"
-    - type: CEZLevelMapRoof
+    - type: CEZLevelRoof
     - type: SunShadow
     - type: MapLight
       ambientLightColor: "#c8fdff"
