+++ Resources/Prototypes/_CE/Procedural/WorldGen/world_configs.yml
@@ -9,7 +9,7 @@
     - type: Gravity
       enabled: true
       inherent: true
-    - type: CEZLevelMapRoof
+    - type: CEZLevelRoof
     - type: SunShadow
       alpha: 0.8
       direction: 2, -0.75
